Working on the config script

+++ src/Core/Config.h
@@ -12,10 +12,10 @@ namespace agio
         static void LoadFromFile(const ConfigLoader& CFGFile);
 
 		// Minimum age before a species is considered to be removed
-		inline static float MinSpeciesAge = 20;
+		inline static int MinSpeciesAge = 20;
 
 		// Number of individuals to consider when computing progress metrics
-		inline static float ProgressMetricsIndividuals = 5;
+		inline static int ProgressMetricsIndividuals = 5;
 
 		// Controls how smooth is the interpolation when computing the progress metric
 		// Lower = more smooth
@@ -26,16 +26,16 @@ namespace agio
 		inline static float ProgressThreshold = 0.005f;
 
 		// Number of consecutive epochs that an species must have a progress lower than the threshold to be removed
-		inline static float SpeciesStagnancyChances = 1e10;//10;
+		inline static int SpeciesStagnancyChances = 1e10;//10;
 
 		// Minimum number of individuals allowed on a species
-		inline static float MinIndividualsPerSpecies = 50;
+		inline static int MinIndividualsPerSpecies = 50;
 
 		// Number of times to try when creating a new morphology
-		inline static float MorphologyTries = 10;
+		inline static int MorphologyTries = 10;
 
 		// Number of replications to do when simulating
-		inline static float SimulationReplications = 10;
+		inline static int SimulationReplications = 10;
 
         inline static float ParameterMutationProb = 0.1f;
         inline static float ParameterDestructiveMutationProb = 0.1f;
